services/api/requirements.txt
Flask_Cors
Flask
requests
python-dateutil
psycopg[binary,pool]
